Create local copies of string properties.
The string values passed to set_property are owned by the caller, so we
need to create new copies of the strings to store in the private data
structure.

The dispose method already had code to free the private strings, but was
missing a check for the en_name property: added one.