Cleaned and updated the time zones SVG.
The SVG map had some issues, and had not been updated in at least a
decade.

Removed layers:
  - s0, which appears to be a scratch layer
  - edit, same
  - tz_export, which looks incomplete and should really be generated
    from the other layers
  - tz_color_codes, ditto
  - bg, since it doesn't contain anything and it was messing up the page
    size

Also removed an extra Iceland that was in layer m1 for some reason.

Added new layers for half and quarter hour zones, except -04:30 which
was already there.

Retraced portions of the map using the August 2013 time zone map from
the CIA World Factbook, which is in the public domain.  Also made some
additional changes not included in the CIA map. The larger changes
include:

  - Updated time zones in Brazil for changes made in 2008 and 2012
  - Updated the Russian time zones, mostly visible on the eastern end
  - Moved Sri Lanka from +06:00 to +05:30, changed in 2006
  - Moved Chile from -04:00 to -03:00 after their change to year-round
    DST in 2015.