Fix memory leaks in tz.c
Use a stack-allocated block for the user_data struct passed to
parse_cities15000, since the struct itself is not needed after
parse_file, and no one can forget to free it if it's not dynamically
allocated the first place.

Use const strings in tz_data_file_get so that the return value does not
need to be duplicated and freed.