Do not use tz_location_get_utc_offset
The function in tz.c to fetch the timezone offset for a given location
sets the TZ environment variable which, besides being potentially
surprising to the rest of the process, is not thread-safe.  Instead,
skip the function entirely and use the glib functions for querying time
zone data.

Increase the required glib version to 2.26 for the GTimeZone functions.