Render the map directly from the SVG.
As high resolution screens become more common, the low resolution of the
map images is becoming more apparent. But the map is already stored as a
vector image, so just use that, instead. Set ids on the layers in the
SVG file to match the timezone offsets, and render the background and
each highlight image from a layer in the SVG.

Drawing from SVG is slower than PNG, so to speed that do the rendering
of the background from SVG to a cairo pattern only when the size of the
widget changes, and add a hash table to cache rendered highlight layers
for a widget given size.

Thanks to Garrett LeSage for the SVG changes.