Improve the location selected when setting the timezone by name.
When setting the timezone from a name (e.g., Australia/Brisbane), try to
find the city that matches the timezone name. That way, in our example,
the pin will appear on Brisbane and not just some random city in eastern
Australia.

In addition, when cc_timezone_map_set_timezone is called for an unknown
timezone, set the offset highlight based on glib's view of timezones
(since this may be a low-population timezone without city data) and
unset the location.