Remove style css "display: none" to fix hidden map
In 0002 patch there is the css style in the svg which is hiding this
svg image. To fix this issue the style value is removed from SVG.

This problem didn't happen till now because of commit:

b2cc9ce181be7dbe90cbea3a57db1e841b0fe826 in librsvg

The commit is setting default style interpretation to css so the code is
processed and it hides the image.

From: Jiri Konecny <jkonecny@redhat.com>