Remove the tests
These tests were intended to check the presence and usability of the
time zone images/layers, which are no longer present. Besides that,
test-clickability had a flaw in that it assumed that the difference
between winter time and summer time is always 60 minutes (this is not
true for Australia/Lord_Howe, which is UTC+10:30 in the winter and
switches to UTC+11 for summer time).

Remove the now unused parts of tz.c. tz.c contains functions for use
within the timezonemap library and are not part of the public API.