Restore some missing islands
The changes to the map made in 5401c28ab0dda465ff3131730614237bb4106720
mistakenly removed several Mediterranean islands in the UTC+1 time zone.
This changes attempts to put them back.

The shapes for the islands were copied from the map in 5401c28~ and
edited to remove the outline and set the fill to same color as the rest
of the map. Their position relative is to each other is from the
previous version of the map, but their position relative to the existing
continents is eyeballed but hopefully close enough to not be visibly
wrong.

This restores Ibiza, Mallorca, Menorca, Corsica, Sardinia, Elba, Sicily,
Malta (main island), Djerba, and 9 shapes off the coast of Croatia that
are harder to pin down as individual islands.